initial commit
create a go-client to defuse_division that connects to a specified
server instance and currently attempts to pull the state of the game.
Currently, main.go will download the game-state and print out the parsed
state, converting from json to go-structs.
You must be running a python server instance of defuse_division. Because
our go-client disconnects and crashes the python server, it's a
convenient idea to run the server in a delayed bash for-loop like so:
while true; ./defuse_division --serveronly; delay 2; done