add initial probability calculations
we calculate probablity that a particular cell contains a mine by
asking a cells' neighbors what they think it's probability is.
Each neighbor returns a % that is equivalent to:
(# of unprobed cells) / (# of mines nearby)
while taking into account known mines.
In addition:
1) add 100ms timeout to reading messages from client. (prevent blocking
code in case we read too many messages. If we see this happening,
re-probing might prompt the server to resend the board state)
2) add render() for solver.Minefield to print out in console
3) initialize MineTouch to -1 if unknown, which allows us to filter out
cells that don't know how many mines they are touching when calculating
probabilities