probe lowest mine-probability cell and repeat
This works by probing the least-likely of primed cells. Primed
cells are those bordering the #'s. Meaning that currently this
solver will proceed in an orderly fashion probing nearby cells
as best as possible.
Then repeat flagging and probing. This continues even after a mine
is accidentally probed. The program exits with an error after the
number of unmarked mines == 0, at which point a nil reference
exception occurs.