add hypothetical scenarios
in the case where we cannot 100% be sure we're probing a non-mine cell,
there are instances where we CAN determine if a cell shouldn't contain
a mine by hypothetically flagging a cell. We then re-calculate
probabilities on the minefield and see if there exists an inconcistency
which would indicate that our hypothetical flag is incorrect. This
allows us to 100% be certain that we can safely probe that cell.