revise hypothetical-mine solving
Hypothetical now correctly identifies when a mine cannot exist at a
location by semi-brute force checking all possibilities until it finds a
cell that CANNOT be a mine. This works fine when there are ~8 revealed
cells. By the time ~11 cells are revealed, this algorithm tends to slow
down incredibly, to the point where this algorithm is no longer
feasible. This will need to be rewritten to be smarter and to include a
timeout.