use linear algebra to solve ambiguous scenarios
theoretically, if we represent the primed-cells to witness # as a
mathematical relationship (such as cell4 + cell3 = 1, meaning that one
of cell4 or cell3 contains a mine), then it should be possible to reduce
some problems down and deduce where mines and flags exist. If we treat
all the witnesses the same and reduce their relationship to their
primed-cell neighbors as a row of #'s, then we can combine them all into
a matrix and perform row operations to calculate which cells contain
mines and which don't. Reduced Row Echelon Form is a reduction of such a
matrix that poses a higher chance of automatically showing which cells
are mines or not. So far, there seems to be a bug. But this holds
promise.