achieve working linear algebra solver
found the problem: witness # (aka the answer-side of the equation) was
overwriting the last coefficient value, rather than being appended to
the last column. I had to compensate by expanding the matrix width to
number of primed-cells + 1. Then the equations worked. Of course, this
is still a basic implementation. It does not yet attempt to further
solve / reduce the matrix but instead relies on some inherent solutions
dropping out from just computing the RREF (Reduced Row Echelon Form).
The next step will be to add a row with negative values in such a way to
cancel out the negative numbers and leave a 0 as the solution
coefficient (which'll solve for more safe cells)