add row recombination and negative no. deduction
row recombination takes a row with a negative coefficient or answer and
adds it to all the other rows of the matrix in the hopes that it'll
simplify a few rows into a logical answer about which is a mine /
non-mine. This has found some good info in test runs.
Similarly, negative deduction acknowledges the fact that the value of
the cell can only be in the set {0, 1}. Meaning that if it's
corresponding coefficient in the row is -1, there's a possibility that
the cell value is 0 (aka not a mine). We can test this by summing up all
the positive valued coefficients. If they equal the answer (last column
of row), then the cell with -1 coefficient must be 0 (non-mine) in
order to satisfy the equation.
With these two added methods, it's very difficult for this solver to
fail. There still exist some ambiguous scenarios... but those are few.
I've noticed that these row-operations are not always consistent in
reporting results. Due to how new info can alter how the Row Reduced
Echelon Form comes out, these algorithms do not always find the same
mines / safe spots they did the turn before. It may be necessary to keep
a permanent list of found mines / safe spots to augment those found by
these matrix methods.
In fact, couple a permanent list with running these matrix operations
for each update (regardless if needed or not) may help deduce
additional mines or safe spots